Add golden sample test for storing transport types kill switch support
This test uses a golden sample to verify that the file reading
results are as expected when the storing transport types feature
is either enabled or disabled. This is to ensure compatibility
when the feature is remotely disabled.

The golden sample is taken from a test device after manually
performing a file download.

Flag: EXEMPT test only
Test: atest ConnectivityCoverageTests:android.net.connectivity.android.net.NetworkStatsCollectionTest \
      --update-device
Test: Manually:
      1. adb shell device_config put tethering store_transport_types 1
      2. Download files
      3. adb shell dumpsys netstats --poll
      4. adb pull /data/misc/apexdata/com.android.tethering/netstats .
      5. Check-in the active netstats_uid* file.

Bug: 419123367
Change-Id: I266d07703f3ad47239d9c6721cdcd417ae4f179f