Wait for Soft AP completion before connecting
The startWifiTethering method is modified to wait for the Soft AP
to be fully started and ready before it returns. Previously, this
method would return as soon as the tethering interface was
active. This could cause race conditions, leading to client
devices attempting to connect to the hotspot before it was fully
operational. The Soft AP is now guaranteed to have successfully
broadcast its information and be ready for connections.

Bug: 417369055
Test: atest CtsConnectivityMultiDevicesTestCases
Flag: EXEMPT bug fix
Change-Id: Ie0420753256bf26c5d44e8c856a87b3c9340a961