bpf_helpers: make *all* eBPF programs return long not int
BPF is a 64-bit architecture, so this makes more sense:
int is 32-bits, while long is 64-bits.

In practice this doesn't do much, but it does make sure
the top 32-bits of the 64-bit R0 register are well defined.

I think this might matter when we try to introduce function calls,
as per Linux Kernel eBPF uapi we have (for example):

  long (*callback_fn)(struct bpf_map *map, const void *key, void *value, void *ctx);
  long bpf_for_each_map_elem(struct bpf_map *map, void *callback_fn, void *callback_ctx, u64 flags);

Test: TreeHugger
Flag: EXEMPT mainline
Signed-off-by: Maciej Żenczykowski <maze@google.com>
Change-Id: I0abad90043eff8f8d99c2c29988dd5ce2fef9927